requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==4.2.18
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1